feat: 그룹 삭제
