+++ src/app/(afterLogin)/group/detail/[groupId]/_lib/fetchDeleteGroup.ts
@@ -1,0 +1,14 @@
+import { fetchWithAuth } from '@/app/(afterLogin)/_lib/FetchWithAuth';
+
+export async function fetchDeleteGroup(groupId: number) {
+  try {
+    const response = await fetchWithAuth(`/groups/${groupId}`, {
+      method: 'DELETE',
+      credentials: 'include',
+      cache: 'no-cache',
+    });
+    return response.data;
+  } catch (error) {
+    console.error(error);
+  }
+}
